Use FStaticMeshLODInfo to reuse the static cached lightmap

+++ Source/LODTests/r_defines.h
@@ -3,4 +3,4 @@
 
 // Unreal seems finnicky with comparisons with #define's? Will have to use const int values instead probably
 
-const int NUM_LODS = 4;
+const uint8_t NUM_LODS = 4;
